requirements.txt
datasets
egises
fire
jieba
nltk
numpy
openai
pandas
PyYAML
rouge_score
safetensors
scipy
sentence_transformers
simtext
spacy
torch
torchmetrics
tqdm
transformers
typer
ujson
vllm
